Split interface and implementation in TextRenderer
Similar that what we did in Batcher class, in this PR we are splitting TextRenderer
with the proper open-gl implementation of the class. In the following PR,
MockTextRenderer will mock TextRenderer class.

A few things:
- We erase TextRenderer.cpp
- UnitTests are missing for all the classes. I will create a bug for that but not address
it in the short time. This is only because I need to Mock TextRenderer.
- We are adding an extra method AddText without output parameters, instead of
having them defaulted to nullptr.
- I'm not changing any extra functionality.

Bug: http://b/228609901.
