+++ src/OrbitGl/TextRenderer.h
@@ -5,54 +5,15 @@
 #ifndef ORBIT_GL_TEXT_RENDERER_H_
 #define ORBIT_GL_TEXT_RENDERER_H_
 
-#include <GteVector.h>
-#include <freetype-gl/mat4.h>
-#include <freetype-gl/texture-atlas.h>
-#include <freetype-gl/texture-font.h>
-#include <freetype-gl/vec234.h>
-#include <glad/glad.h>
-#include <stddef.h>
-#include <stdint.h>
-
-#include <map>
-#include <unordered_map>
-#include <vector>
-
-#include "CoreMath.h"
-#include "PickingManager.h"
-#include "PrimitiveAssembler.h"
 #include "TextRendererInterface.h"
 #include "Viewport.h"
-
-namespace ftgl {
-struct vertex_buffer_t;
-struct texture_font_t;
-}  // namespace ftgl
 
 namespace orbit_gl {
 
 class TextRenderer : public TextRendererInterface {
  public:
-  explicit TextRenderer();
-  ~TextRenderer();
-
-  void Init() override;
-  void Clear() override;
+  explicit TextRenderer() {}
   void SetViewport(Viewport* viewport) { viewport_ = viewport; }
-
-  void RenderLayer(float layer) override;
-  void RenderDebug(PrimitiveAssembler* primitive_assembler) override;
-  [[nodiscard]] std::vector<float> GetLayers() const override;
-
-  void AddText(const char* text, float x, float y, float z, TextFormatting formatting,
-               Vec2* out_text_pos = nullptr, Vec2* out_text_size = nullptr) override;
-
-  float AddTextTrailingCharsPrioritized(const char* text, float x, float y, float z,
-                                        TextFormatting formatting,
-                                        size_t trailing_chars_length) override;
-
-  [[nodiscard]] float GetStringWidth(const char* text, uint32_t font_size) override;
-  [[nodiscard]] float GetStringHeight(const char* text, uint32_t font_size) override;
 
   void PushTranslation(float x, float y, float z = 0.f) { translations_.PushTranslation(x, y, z); }
   void PopTranslation() { translations_.PopTranslation(); }
@@ -60,46 +21,11 @@ class TextRenderer : public TextRendererInterface {
   static void SetDrawOutline(bool value) { draw_outline_ = value; }
 
  protected:
-  void AddTextInternal(const char* text, ftgl::vec2* pen, const TextFormatting& formatting, float z,
-                       ftgl::vec2* out_text_pos = nullptr, ftgl::vec2* out_text_size = nullptr);
-
-  [[nodiscard]] int GetStringWidthScreenSpace(const char* text, uint32_t font_size);
-  [[nodiscard]] int GetStringHeightScreenSpace(const char* text, uint32_t font_size);
-  [[nodiscard]] ftgl::texture_font_t* GetFont(uint32_t size);
-  [[nodiscard]] ftgl::texture_glyph_t* MaybeLoadAndGetGlyph(ftgl::texture_font_t* self,
-                                                            const char* character);
-
-  void DrawOutline(PrimitiveAssembler* primitive_assembler, ftgl::vertex_buffer_t* buffer);
-
- private:
-  ftgl::texture_atlas_t* texture_atlas_;
-  // Indicates when a change to the texture atlas occurred so that we have to reupload the
-  // texture data. Only freetype-gl's texture_font_load_glyph modifies the texture atlas,
-  // so we need to set this to true when and only when we call that function.
-  bool texture_atlas_changed_;
-  std::unordered_map<float, ftgl::vertex_buffer_t*> vertex_buffers_by_layer_;
-  std::map<uint32_t, ftgl::texture_font_t*> fonts_by_size_;
-  Viewport* viewport_;
-  GLuint shader_;
-  ftgl::mat4 model_;
-  ftgl::mat4 view_;
-  ftgl::mat4 projection_;
-  ftgl::vec2 pen_;
-  bool initialized_;
-  static bool draw_outline_;
+  Viewport* viewport_ = nullptr;
+  inline static bool draw_outline_ = false;
 
   TranslationStack translations_;
 };
-
-inline ftgl::vec4 ColorToVec4(const Color& color) {
-  const float coeff = 1.f / 255.f;
-  ftgl::vec4 vec;
-  vec.r = color[0] * coeff;
-  vec.g = color[1] * coeff;
-  vec.b = color[2] * coeff;
-  vec.a = color[3] * coeff;
-  return vec;
-}
 
 }  // namespace orbit_gl
